wip: distinct camera component, to help with growing complexity
