+++ frontend/iou/src/views/WorkspacePage/WorkspacePage.tsx
@@ -3,35 +3,18 @@ import {Button} from "@/components/ui/button.tsx";
 import { LogOut } from "lucide-react"
 import WorkspaceMenubar from "@/components/widgets/workspace/WorkspaceMenubar.tsx";
 import ContextSidebar from "@/components/widgets/workspace/ContextSidebar.tsx";
-import {CameraControls, Grid, OrthographicCamera, PerspectiveCamera} from "@react-three/drei";
+import {Grid} from "@react-three/drei";
 import useDimensions from "@/hooks/workspace/useDimensions.ts";
 import useShapeUUIDs from "@/hooks/workspace/useShapeUUIDs.tsx";
 import ShapeWidget from "@/components/three/shape/ShapeWidget.tsx";
 import WorkspaceTitle from "@/components/widgets/workspace/WorkspaceTitle.tsx";
 import WorkspaceActionListener from "@/components/widgets/workspace/WorkspaceActionListener.tsx";
 import VertexControls from "@/components/three/VertexControls.tsx";
-import useCameraInteraction from "@/hooks/workspace/useCameraInteraction.ts";
 
 export default function WorkspacePage() {
   const [dimensions, setDimensions] = useDimensions();
 
   const shapeUUIDs = useShapeUUIDs();
-  const cameraInteraction = useCameraInteraction();
-
-  const camera =dimensions === "2d" ? (
-    <OrthographicCamera
-      makeDefault
-      zoom={200}
-      position={[0, 1.5, 100]}
-    />
-  ) : (
-    <>
-      <PerspectiveCamera
-        makeDefault
-        position={[0, 1.5, 5]}
-      />
-    </>
-  );
 
   // These are all the JSX elements used as an overlay on top of the 3d/2d view
   const overlay = (
@@ -72,8 +55,6 @@ export default function WorkspacePage() {
       className="w-screen h-screen overflow-clip bg-secondary"
       overlay={overlay}
     >
-      {camera}
-      <CameraControls enabled={cameraInteraction === undefined}></CameraControls>
 
       <Grid
         infiniteGrid       // <- key flag
